add dry run functionality as development help
The dry run feature runs the entire flash process together with the real
target but does not execute the flash write and flash erase commands.

As a result return values from the target do not match the original process
at some points but it helps to see what kind to data and checksums are
exchanged between pcanflash and the target.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>